add: 게시글 카드 클릭 시 해당 게시글 모달 창이 뜨는 UI 작업
- BoardCard 컴포넌트화
- 타입 '?' 지정
