+++ src/pages/NoticeBoard/NoticeBoardModal/components/BoardModalTyping.tsx
