+++ src/types/type.ts
@@ -39,6 +39,19 @@ export interface ChatModalProp {
 
 export type idProp = number | undefined;
 
+export interface BoardDataProp {
+  id?: number;
+  title?: string;
+  image?: string;
+  date?: string;
+  writer?: string;
+  like?: number;
+}
+
 export interface BoardModalProp {
-  clickCard: () => void;
+  clickCard?: () => void;
+  title?: string;
+  date?: string;
+  image?: string;
+  like?: number;
 }
